Updates in logout and login so they can handle the password. Also updated the clear method in SQlGameDAO. About to do some tests.

+++ server/src/main/java/servicefiles/LogoutService.java
@@ -14,6 +14,13 @@ public class LogoutService {
         if (authToken == null || authToken.isBlank()) {
             throw new DataAccessException("No hay Token");
         }
+
+
+        if (authDAO.getAuth(authToken) == null) {
+            throw new DataAccessException("Token not found");
+        }
+
         authDAO.deleteAuth(authToken);
     }
+
 }
